refactor: reveiw 관련 api 정리
- 타입, api 통합
- 타입 재사용
- 잘못된 네이밍(props -> params) 수정
