+++ src/entities/gathering-detail/ui/ReviewList.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useTranslations } from 'next-intl';
-import { useGatheringReviewList } from '@/entities/review/hooks/useGatheringReviewList';
+import { useGetGatheringReviewList } from '@/entities/review/api/queries';
 import { ReviewCard } from '@/entities/review/ui/ReviewCard';
 import { Pagination } from '@/shared/ui/pagination';
 
@@ -13,7 +13,7 @@ export const ReviewList = ({ id }: { id: number }) => {
   const [page, setPage] = useState(1);
   const offset = (page - 1) * PAGE_SIZE;
 
-  const { data, isLoading } = useGatheringReviewList(id, {
+  const { data, isLoading } = useGetGatheringReviewList(id, {
     offset,
     limit: PAGE_SIZE,
     sortBy: 'createdAt',
