+++ src/entities/review/api/queries.ts
@@ -1,20 +1,33 @@
+import { GatheringType } from '@/entities/gathering/model/types';
+import { CreateReviewPayload } from '@/entities/review/model/schemas';
+import {
+  CreateReviewResponse,
+  ReviewFilterParams,
+  ReviewListResponse,
+  ReviewScoreItem,
+} from '@/entities/review/model/type';
 import { QUERY_KEYS } from '@/shared/api';
-import { useSuspenseInfiniteQuery, useSuspenseQuery } from '@tanstack/react-query';
-import { ReviewFilterProps, ReviewListResponse, ReviewScoreItem, ReviewType } from '../model/type';
-import { getReviewList, getReviewScore } from './services';
+import { ApiError } from '@/shared/api';
+import {
+  useMutation,
+  useQuery,
+  useQueryClient,
+  useSuspenseInfiniteQuery,
+  useSuspenseQuery,
+} from '@tanstack/react-query';
+import { keepPreviousData } from '@tanstack/react-query';
+import { createReview, getGatheringReviewList, getReviewList, getReviewScore } from './services';
 
 // 리뷰 점수 조회
-export const useGetReviewScore = (type?: ReviewType) => {
+export const useGetReviewScore = (type?: GatheringType) => {
   return useSuspenseQuery<ReviewScoreItem[], Error>({
     queryKey: QUERY_KEYS.review.scores({ type }),
     queryFn: () => getReviewScore({ type }),
   });
 };
 
-//
-
-// 리뷰리스트 무한스크롤
-export const useGetReviewListInfinite = (filters: ReviewFilterProps) => {
+// 리뷰 리스트 무한스크롤
+export const useGetReviewListInfinite = (filters: ReviewFilterParams) => {
   return useSuspenseInfiniteQuery<ReviewListResponse>({
     queryKey: QUERY_KEYS.review.list(filters),
     queryFn: ({ pageParam = 0 }) => {
@@ -29,5 +42,38 @@ export const useGetReviewListInfinite = (filters: ReviewFilterProps) => {
       if (currentPage <= totalPages - 1) return currentPage;
       return undefined;
     },
+  });
+};
+
+// 리뷰 조회
+export const useGetReviews = (params: ReviewFilterParams, options?: { enabled?: boolean }) => {
+  return useQuery<ReviewListResponse>({
+    queryKey: QUERY_KEYS.review.list(params),
+    queryFn: () => getReviewList(params),
+    enabled: options?.enabled ?? true,
+  });
+};
+
+// 리뷰 생성
+export const useCreateReview = () => {
+  const queryClient = useQueryClient();
+  return useMutation<CreateReviewResponse, ApiError, CreateReviewPayload>({
+    mutationFn: (payload) => createReview(payload),
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: QUERY_KEYS.review.base });
+      queryClient.invalidateQueries({ queryKey: QUERY_KEYS.gathering.base });
+    },
+  });
+};
+
+// 특정 모임 리뷰 조회
+export const useGetGatheringReviewList = (
+  gatheringId: number,
+  params: Omit<ReviewFilterParams, 'gatheringId'>,
+) => {
+  return useQuery<ReviewListResponse>({
+    queryKey: QUERY_KEYS.gathering.reviews(gatheringId, params),
+    queryFn: () => getGatheringReviewList(gatheringId, params),
+    placeholderData: keepPreviousData,
   });
 };
