+++ src/entities/review/api/services.ts
@@ -1,28 +1,37 @@
 import { httpClient } from '@/shared/api';
 import { API_ENDPOINTS } from '@/shared/config';
+import { CreateReviewPayload } from '../model/schemas';
 import {
-  ReviewFilterProps,
+  CreateReviewResponse,
+  ReviewFilterParams,
   ReviewListResponse,
-  ReviewScoreProps,
+  ReviewScoreParams,
   ReviewScoreResponse,
 } from '../model/type';
 
-// 전체목록 조회
-export const getReviewList = (params: ReviewFilterProps): Promise<ReviewListResponse> => {
+// 리뷰 조회
+export const getReviewList = (params: ReviewFilterParams): Promise<ReviewListResponse> => {
   return httpClient.get(`${API_ENDPOINTS.REVIEWS.LIST}`, { params });
 };
 
-//스토어 점수 조회
-export const getReviewScore = (params: ReviewScoreProps = {}): Promise<ReviewScoreResponse> => {
+// 리뷰 점수 조회
+export const getReviewScore = (params: ReviewScoreParams = {}): Promise<ReviewScoreResponse> => {
   return httpClient.get(`${API_ENDPOINTS.REVIEWS.SCORES}`, { params });
 };
 
 //특정 모임 리뷰 조회
 export const getGatheringReviewList = (
   gatheringId: number,
-  params: Omit<ReviewFilterProps, 'gatheringId'> = {},
+  params: Omit<ReviewFilterParams, 'gatheringId'> = {},
 ): Promise<ReviewListResponse> => {
   return httpClient.get(`${API_ENDPOINTS.REVIEWS.LIST}`, {
     params: { ...params, gatheringId },
+  });
+};
+
+// 리뷰 생성
+export const createReview = (payload: CreateReviewPayload): Promise<CreateReviewResponse> => {
+  return httpClient.post(`${API_ENDPOINTS.REVIEWS.CREATE}`, payload, {
+    authRequired: true,
   });
 };
